Update bead_detector_v0.py
I have added a local mean speed method.  After collecting data it divides the frame into a grid defined by 'size', and finds the mean velocity in each of this areas. These means are then used as better 'guesses' for locating the position of the bead in the next frame.  It is an improvement ont he global mean (perhaps!).  To return to a global mean set the size=(1,1).